+++ apps/docs/src/content/03-components/data-visualisation/donut-chart/examples/decimal.tsx
@@ -1,0 +1,7 @@
+import { DonutChart } from "@mittwald/flow-react-components";
+
+<DonutChart
+  formatOptions={{ style: "decimal" }}
+  maxValue={200}
+  value={135}
+/>;