+++ apps/docs/src/content/03-components/data-visualisation/donut-chart/examples/default.tsx
@@ -1,0 +1,3 @@
+import { DonutChart } from "@mittwald/flow-react-components";
+
+<DonutChart value={30} />;