+++ apps/docs/src/content/03-components/data-visualisation/donut-chart/examples/unit.tsx
@@ -1,0 +1,7 @@
+import { DonutChart } from "@mittwald/flow-react-components";
+
+<DonutChart
+  formatOptions={{ style: "unit", unit: "gigabyte" }}
+  maxValue={50}
+  value={12}
+/>;