+++ packages/components/src/components/DonutChart/view.ts
@@ -1,0 +1,10 @@
+/* prettier-ignore */
+/* This file is auto-generated with the remote-components-generator */
+import type { DonutChart } from "./DonutChart";
+import type { ViewComponent } from "@/lib/viewComponentContext";
+
+declare global {
+  interface FlowViewComponents {
+    DonutChart: ViewComponent<typeof DonutChart>;
+  }
+}